requirements.txt
streamlit==1.27.0
langchain==0.0.305
langchain-community==0.0.13
langchain-google-genai==0.0.5
google-generativeai==0.3.1
python-dotenv==1.0.0
SpeechRecognition==3.10.0
faiss-cpu==1.7.4
requests==2.31.0
pyaudio==0.2.13
pydub==0.25.1 